+++ src/custom/pages/About/index.tsx
@@ -1,26 +1,11 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
+import useMarkdown from '../../hooks/useMarkdown'
 import { Title, Content, AppBodyMod } from './About.styled'
 import markdownContent from './about.md'
 import ReactMarkdown from 'react-markdown'
 
 export default function About() {
-  const [content, setContent] = useState('')
-
-  useEffect(() => {
-    const fetchContent = async () => {
-      await fetch(markdownContent)
-        .then(res => res.text())
-        .then(text => {
-          setContent(text)
-        })
-        .catch(error => {
-          console.log('Error fetching markdown content: ', error)
-          return null
-        })
-    }
-
-    fetchContent()
-  }, [])
+  const content = useMarkdown(markdownContent)
 
   return (
     <AppBodyMod>
