+++ src/custom/pages/App/index.tsx
@@ -4,6 +4,7 @@ import styled from 'styled-components'
 import { RedirectPathToSwapOnly, RedirectToSwap } from 'pages/Swap/redirects'
 import { Route, Switch } from 'react-router-dom'
 import Swap from 'pages/Swap'
+import About from 'pages/About'
 
 export const Wrapper = styled(AppMod)``
 
@@ -14,6 +15,7 @@ export default function App() {
         <Route exact strict path="/swap" component={Swap} />
         <Route exact strict path="/swap/:outputCurrency" component={RedirectToSwap} />
         <Route exact strict path="/send" component={RedirectPathToSwapOnly} />
+        <Route exact strict path="/about" component={About} />
         <Route component={RedirectPathToSwapOnly} />
       </Switch>
     </Wrapper>
